backend/requirements.txt
django~=4.2.25
django-admin-autocomplete-filter
django-autocomplete-light
django-cities
django-cors-headers
django-cron
django-debug-toolbar
django-extensions
django-filter
django-jazzmin
django-ordered-model==3.7.4
djangorestframework==3.14.0
djangorestframework-gis==1.0
importlib-metadata==6.8.0
Markdown==3.4.4
pdfkit==1.0.0
psycopg2-binary==2.9.7
PyJWT==2.8.0
pyOpenSSL==23.2.0
qrcode==7.4.2
requests-pkcs12==1.18
uWSGI==2.0.22
pytest
pytest-asyncio
pytest-mock